fastApiTestingProject/pyproject.toml
[tool.poetry]
name = "fastapitestingproject"
version = "0.1.0"
description = ""
authors = ["daniilkimstach <daniilkimstachp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.1"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
fastapi = "^0.110.2"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
loguru = "^0.7.2"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
